Fix behavior of `filled` macro for Params schemas
This adds a customized handling of empty strings in case of Params
schemas. They are now configured to use `filter(:filled?)` which is
applied before coercing an empty string to the expected type.

As a result you get better errors when empty strings are passed as
values and we can catch that before coercion, which should result
in better performance.

  * Add DSL#[] to ease accessing key macros
  * Add `PrimitiveInferrer`

Fixes #134