Allow unexpected keys to cause result failure
The success of a schema result depended on `results` being empty. This
used to be correct as any steps which failed would add to `results`.

However, since the addition of the `key_validator` this is no longer
true. The `key_validator` adds errors directly to the `result_ast`
(which is then used to generate error messages) but doesn't change the
contents of `results`.

In order for unexpected keys to caused `.success?` to return `false` we
simply need to check whether `result_ast` is empty, rather than
`results`.

Fixes #297